diffusionx::random::uniform

Function inclusive_range_rands

pub fn inclusive_range_rands<T>(
    range: RangeInclusive<T>,
    n: usize,
) -> XResult<Vec<T>>
where
    T: SampleUniform + Send + Sync,
    Uniform<T>: Copy,
    <T as SampleUniform>::Sampler: Send + Sync,

Generate a vector of random numbers from an inclusive range

This function generates a vector of random numbers from an inclusive range using the Uniform distribution.

Returns

A vector of T values representing the generated random numbers.

Example

use diffusionx::random::uniform::inclusive_range_rands;
let randoms = inclusive_range_rands(0..=10, 10).unwrap();
assert_eq!(randoms.len(), 10);
assert!(randoms.iter().all(|x| (0..=10).contains(x)));