diffusionx

Module utils

Utils

This module provides utility functions.

Functions

cumsum: Calculate the cumulative sum of a vector.

gamma: Calculate the gamma function of a number.

gammaf: Calculate the gamma function of a number (f32 version).

approx_eq: Check if two numbers are approximately equal.

float_eq: Check if two numbers are equal.

eval_poly: Evaluate a polynomial.

sinpi: Calculate the sine of pi times a number.

cospi: Calculate the cosine of pi times a number.

sincospi: Calculate the sine and cosine of pi times a number.

tanpi: Calculate the tangent of pi times a number.

minmax: Find the minimum and maximum values in a vector.

calculate_stats: Calculate the mean and variance of a vector, using in test.

calculate_int_stats: Calculate the mean and variance of an integer vector, using in test.

calculate_bool_mean: Calculate the mean of a boolean vector, using in test.

Functions

approx_eq

Check if two floating numbers are approximately equal within a tolerance

cospi

Calculate cos(pi * x)

cumsum

Calculate the cumulative sum of a vector

eval_poly

Evaluate a polynomial using the Horner method

float_eq

Check if two floating numbers are equal within the f64 precision

gamma

Gamma function

gammaf

Gamma function for f32

minmax

find max value and min value in a &[f64]

sincospi

Calculate sin(pi * x) and cos(pi * x)

sinpi

Calculate sin(pi * x)

tanpi

Calculate tan(pi * x)

tgamma⚠

tgammaf⚠